requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.16
django-ckeditor==5.9.0
django-debug-toolbar==3.7.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-phonenumber-field==7.3.0
django-redis==5.2.0
django-taggit==1.2.0
django-user-visit==0.4
gunicorn==20.0.4
importlib-metadata==5.0.0
packaging==21.3
phonenumbers==8.13.30
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyparsing==3.0.9
pytz==2019.3
redis==4.3.4
sqlparse==0.3.1
typing_extensions==4.4.0
ua-parser==0.10.0
user-agents==2.2.0
whitenoise==5.0.1
wrapt==1.14.1
zipp==3.10.0
